Move HATEOAS link creation to AbsenceDto
